backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
groq==0.9.0
httpx==0.27.2
PyMuPDF==1.24.11
reportlab==4.2.2
Pillow==10.4.0
